academic-audit-system-analyzer/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==3.2.16
dnspython==2.7.0
idna==3.10
joblib==1.4.2
kafka-python==2.0.2
lxml==5.3.2
nltk==3.8.1
psycopg2-binary==2.9.6
pymongo==4.3.3
PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2025.2
regex==2024.11.6
requests==2.31.0
soupsieve==2.6
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
